requirements.txt
Django==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.29.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-ckeditor==6.7.3
Pillow==12.0.0
django-imagekit==6.0.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
zeep>=4.3.0

